# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MediSmart API",
    "description": "\n## 🏥 MediSmart — Hệ thống tra cứu thuốc & tư vấn sức khỏe thông minh\n\n### 🔑 Authentication\n- Lấy token: `POST /api/v1/auth/login`\n- Gắn vào header: `Authorization: Bearer {access_token}`\n- Access token hết hạn sau **30 phút** → dùng `POST /api/v1/auth/refresh`\n- Logout → token bị thu hồi ngay lập tức (Redis blacklist)\n\n### 👥 Phân quyền\n| Role  | Quyền truy cập |\n|-------|----------------|\n| User  | Hồ sơ cá nhân, tra cứu thuốc, chatbot, lịch sử |\n| Admin | Tất cả của User + quản lý dữ liệu + system logs |\n\n### 📄 Response Format\nTất cả responses đều theo format:\n```json\n{\n  \"success\": true,\n  \"message\": \"Thành công\",\n  \"data\": { ... }\n}\n```\n\n### ⚠️ Rate Limiting\n- Login: 5 lần / 15 phút / IP\n- Gửi lại email xác thực: 1 lần / 5 phút / user\n- Reset mật khẩu: 1 lần / 5 phút / user\n    ",
    "contact": {
      "name": "MediSmart Support",
      "email": "support@medismart.vn"
    },
    "license": {
      "name": "Private"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/drugs": {
      "get": {
        "tags": [
          "💊 Thuốc"
        ],
        "summary": "Danh sách thuốc",
        "description": "Tìm kiếm theo tên, mã DB hoặc mã ATC. Public — không cần đăng nhập.",
        "operationId": "list_drugs_api_v1_drugs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Size"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Tìm theo tên / mã DB / ATC",
              "title": "Search"
            },
            "description": "Tìm theo tên / mã DB / ATC"
          },
          {
            "name": "dosage_form",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Lọc theo dạng bào chế",
              "title": "Dosage Form"
            },
            "description": "Lọc theo dạng bào chế"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drugs/{drug_id}": {
      "get": {
        "tags": [
          "💊 Thuốc"
        ],
        "summary": "Chi tiết thuốc",
        "operationId": "get_drug_api_v1_drugs__drug_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "drug_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Drug Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrugDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drugs/{drug_id}/interactions": {
      "get": {
        "tags": [
          "💊 Thuốc"
        ],
        "summary": "Tương tác của một thuốc",
        "description": "Liệt kê tất cả cặp tương tác liên quan đến thuốc này.",
        "operationId": "get_drug_interactions_api_v1_drugs__drug_id__interactions_get",
        "parameters": [
          {
            "name": "drug_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Drug Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "🔧 Admin"
        ],
        "summary": "Danh sách người dùng",
        "operationId": "list_users_api_v1_admin_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Size"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}": {
      "get": {
        "tags": [
          "🔧 Admin"
        ],
        "summary": "Chi tiết người dùng",
        "operationId": "get_user_api_v1_admin_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "🔧 Admin"
        ],
        "summary": "Cập nhật thông tin người dùng",
        "operationId": "update_user_api_v1_admin_users__user_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/toggle-active": {
      "patch": {
        "tags": [
          "🔧 Admin"
        ],
        "summary": "Kích hoạt / Vô hiệu hóa tài khoản",
        "operationId": "toggle_active_api_v1_admin_users__user_id__toggle_active_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/drugs": {
      "post": {
        "tags": [
          "🔧 Admin"
        ],
        "summary": "Thêm thuốc mới",
        "operationId": "create_drug_api_v1_admin_drugs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminDrugCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/drugs/{drug_id}": {
      "put": {
        "tags": [
          "🔧 Admin"
        ],
        "summary": "Cập nhật thông tin thuốc",
        "operationId": "update_drug_api_v1_admin_drugs__drug_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "drug_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Drug Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminDrugUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "🔧 Admin"
        ],
        "summary": "Xóa thuốc",
        "operationId": "delete_drug_api_v1_admin_drugs__drug_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "drug_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Drug Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/drugs/{drug_id}/products": {
      "post": {
        "tags": [
          "🔧 Admin"
        ],
        "summary": "Thêm thông tin sản phẩm",
        "operationId": "add_product_api_v1_admin_drugs__drug_id__products_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "drug_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Drug Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminProductCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/drugs/{drug_id}/warnings": {
      "post": {
        "tags": [
          "🔧 Admin"
        ],
        "summary": "Thêm cảnh báo thuốc",
        "operationId": "add_warning_api_v1_admin_drugs__drug_id__warnings_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "drug_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Drug Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminWarningCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/drugs/{drug_id}/warnings/{warning_id}": {
      "delete": {
        "tags": [
          "🔧 Admin"
        ],
        "summary": "Xóa cảnh báo thuốc",
        "operationId": "delete_warning_api_v1_admin_drugs__drug_id__warnings__warning_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "drug_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Drug Id"
            }
          },
          {
            "name": "warning_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Warning Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/logs/system/export": {
      "get": {
        "tags": [
          "🔧 Admin"
        ],
        "summary": "Xuất system logs ra Excel",
        "operationId": "export_system_logs_api_v1_admin_logs_system_export_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminWarningCreate": {
        "properties": {
          "warning_text": {
            "type": "string",
            "title": "Warning Text"
          }
        },
        "type": "object",
        "required": [
          "warning_text"
        ],
        "title": "AdminWarningCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}